refactor: Replace `Once` with now stable `OnceLock`
This refactor splits a unit test into two integration tests as it's
no longer possible to overwrite `SERDE_HUMAN_READABLE` after it has
been set. By splitting it into two integration tests, they don't
share global state and can thus modify it for the test.

I've also removed the unit tests that test that the `Serializer`/
`Deserializer` correctly use `SERDE_HUMAN_READABLE` for their
`is_human_readable` implementation as these structs are not accessible
from a integration test. I've kept the half of the tests that test
for `SERDE_HUMAN_READABLE == true`. If this implementation is ever
changed it will also affect the integration tests so it's still being
tested.
