+++ avro/Cargo.toml
@@ -88,7 +88,6 @@ criterion = { default-features = false, version = "0.7.0" }
 hex-literal = { default-features = false, version = "1.0.0" }
 md-5 = { default-features = false, version = "0.10.6" }
 pretty_assertions = { workspace = true }
-serial_test = "3.2.0"
 sha2 = { default-features = false, version = "0.10.9" }
 paste = { default-features = false, version = "1.0.15" }
 rstest = { default-features = false, version = "0.26.1" }
